make stage -p show the number of changes in a file

+++ lib/got_lib_diff.h
@@ -93,7 +93,7 @@ struct got_diff_change {
 };
 
 struct got_diff_changes {
-	size_t nchanges;
+	int nchanges;
 	SIMPLEQ_HEAD(, got_diff_change) entries;
 };
